✨ feat(account): add account pages and update header links
