+++ resources/components/AppSidebar.vue
@@ -36,7 +36,7 @@ const footerNavItems: NavItem[] = [
             <SidebarMenu>
                 <SidebarMenuItem>
                     <SidebarMenuButton size="lg" as-child>
-                        <Link :href="route('dashboard')">
+                        <Link :href="route('dashboard.index')">
                             <AppLogo />
                         </Link>
                     </SidebarMenuButton>
