+++ resources/views/account/Index.vue
@@ -1,0 +1,33 @@
+<script setup lang="ts">
+import PlaceholderPattern from '@/components/PlaceholderPattern.vue';
+import { type BreadcrumbItem } from '@/types';
+import { Head } from '@inertiajs/vue3';
+
+const breadcrumbs: BreadcrumbItem[] = [
+    {
+        title: 'My Account',
+        href: '/account',
+    },
+];
+</script>
+
+<template>
+    <Head title="Dashboard" />
+
+    <div class="flex h-full flex-1 flex-col gap-4 overflow-x-auto rounded-xl p-4">
+        <div class="grid auto-rows-min gap-4 md:grid-cols-3">
+            <div class="relative aspect-video overflow-hidden rounded-xl border border-sidebar-border/70 dark:border-sidebar-border">
+                <PlaceholderPattern />
+            </div>
+            <div class="relative aspect-video overflow-hidden rounded-xl border border-sidebar-border/70 dark:border-sidebar-border">
+                <PlaceholderPattern />
+            </div>
+            <div class="relative aspect-video overflow-hidden rounded-xl border border-sidebar-border/70 dark:border-sidebar-border">
+                <PlaceholderPattern />
+            </div>
+        </div>
+        <div class="relative min-h-[100vh] flex-1 rounded-xl border border-sidebar-border/70 md:min-h-min dark:border-sidebar-border">
+            <PlaceholderPattern />
+        </div>
+    </div>
+</template>
