+++ resources/views/account/settings/Password.vue
@@ -1,0 +1,98 @@
+<script setup lang="ts">
+import InputError from '@/components/InputError.vue';
+import SettingsLayout from '@/layouts/settings/Layout.vue';
+import { Form, Head } from '@inertiajs/vue3';
+import { ref } from 'vue';
+
+import HeadingSmall from '@/components/HeadingSmall.vue';
+import { Button } from '@/components/ui/button';
+import { Input } from '@/components/ui/input';
+import { Label } from '@/components/ui/label';
+import { type BreadcrumbItem } from '@/types';
+
+const breadcrumbItems: BreadcrumbItem[] = [
+    {
+        title: 'Password settings',
+        href: '/settings/password',
+    },
+];
+
+const passwordInput = ref<HTMLInputElement | null>(null);
+const currentPasswordInput = ref<HTMLInputElement | null>(null);
+</script>
+
+<template>
+    <Head title="Password settings" />
+
+    <SettingsLayout>
+        <div class="space-y-6">
+            <HeadingSmall title="Update password" description="Ensure your account is using a long, random password to stay secure" />
+
+            <Form
+                method="put"
+                :action="route('account.password.update')"
+                :options="{
+                    preserveScroll: true,
+                }"
+                reset-on-success
+                :reset-on-error="['password', 'password_confirmation', 'current_password']"
+                class="space-y-6"
+                v-slot="{ errors, processing, recentlySuccessful }"
+            >
+                <div class="grid gap-2">
+                    <Label for="current_password">Current password</Label>
+                    <Input
+                        id="current_password"
+                        ref="currentPasswordInput"
+                        name="current_password"
+                        type="password"
+                        class="mt-1 block w-full"
+                        autocomplete="current-password"
+                        placeholder="Current password"
+                    />
+                    <InputError :message="errors.current_password" />
+                </div>
+
+                <div class="grid gap-2">
+                    <Label for="password">New password</Label>
+                    <Input
+                        id="password"
+                        ref="passwordInput"
+                        name="password"
+                        type="password"
+                        class="mt-1 block w-full"
+                        autocomplete="new-password"
+                        placeholder="New password"
+                    />
+                    <InputError :message="errors.password" />
+                </div>
+
+                <div class="grid gap-2">
+                    <Label for="password_confirmation">Confirm password</Label>
+                    <Input
+                        id="password_confirmation"
+                        name="password_confirmation"
+                        type="password"
+                        class="mt-1 block w-full"
+                        autocomplete="new-password"
+                        placeholder="Confirm password"
+                    />
+                    <InputError :message="errors.password_confirmation" />
+                </div>
+
+                <div class="flex items-center gap-4">
+                    <Button :disabled="processing">Save password</Button>
+
+                    <Transition
+                        enter-active-class="transition ease-in-out"
+                        enter-from-class="opacity-0"
+                        leave-active-class="transition ease-in-out"
+                        leave-to-class="opacity-0"
+                    >
+                        <p v-show="recentlySuccessful" class="text-sm text-neutral-600">Saved.</p>
+                    </Transition>
+                </div>
+            </Form>
+        </div>
+    </SettingsLayout>
+</template>
