+++ resources/views/account/settings/Profile.vue
@@ -1,0 +1,103 @@
+<script setup lang="ts">
+import { Form, Head, Link, usePage } from '@inertiajs/vue3';
+
+import DeleteUser from '@/components/DeleteUser.vue';
+import HeadingSmall from '@/components/HeadingSmall.vue';
+import InputError from '@/components/InputError.vue';
+import { Button } from '@/components/ui/button';
+import { Input } from '@/components/ui/input';
+import { Label } from '@/components/ui/label';
+import SettingsLayout from '@/layouts/settings/Layout.vue';
+import { type BreadcrumbItem, type User } from '@/types';
+
+interface Props {
+    mustVerifyEmail: boolean;
+    status?: string;
+}
+
+defineProps<Props>();
+
+const breadcrumbItems: BreadcrumbItem[] = [
+    {
+        title: 'Profile settings',
+        href: '/settings/profile',
+    },
+];
+
+const page = usePage();
+const user = page.props.auth.user as User;
+</script>
+
+<template>
+    <Head title="Profile settings" />
+
+    <SettingsLayout>
+        <div class="flex flex-col space-y-6">
+            <HeadingSmall title="Profile information" description="Update your name and email address" />
+
+            <Form method="patch" :action="route('account.profile.update')" class="space-y-6" v-slot="{ errors, processing, recentlySuccessful }">
+                <div class="grid gap-2">
+                    <Label for="name">Name</Label>
+                    <Input
+                        id="name"
+                        class="mt-1 block w-full"
+                        name="name"
+                        :default-value="user.name"
+                        required
+                        autocomplete="name"
+                        placeholder="Full name"
+                    />
+                    <InputError class="mt-2" :message="errors.name" />
+                </div>
+
+                <div class="grid gap-2">
+                    <Label for="email">Email address</Label>
+                    <Input
+                        id="email"
+                        type="email"
+                        class="mt-1 block w-full"
+                        name="email"
+                        :default-value="user.email"
+                        required
+                        autocomplete="username"
+                        placeholder="Email address"
+                    />
+                    <InputError class="mt-2" :message="errors.email" />
+                </div>
+
+                <div v-if="mustVerifyEmail && !user.email_verified_at">
+                    <p class="-mt-4 text-sm text-muted-foreground">
+                        Your email address is unverified.
+                        <Link
+                            :href="route('verification.send')"
+                            method="post"
+                            as="button"
+                            class="text-foreground underline decoration-neutral-300 underline-offset-4 transition-colors duration-300 ease-out hover:decoration-current! dark:decoration-neutral-500"
+                        >
+                            Click here to resend the verification email.
+                        </Link>
+                    </p>
+
+                    <div v-if="status === 'verification-link-sent'" class="mt-2 text-sm font-medium text-green-600">
+                        A new verification link has been sent to your email address.
+                    </div>
+                </div>
+
+                <div class="flex items-center gap-4">
+                    <Button :disabled="processing">Save</Button>
+
+                    <Transition
+                        enter-active-class="transition ease-in-out"
+                        enter-from-class="opacity-0"
+                        leave-active-class="transition ease-in-out"
+                        leave-to-class="opacity-0"
+                    >
+                        <p v-show="recentlySuccessful" class="text-sm text-neutral-600">Saved.</p>
+                    </Transition>
+                </div>
+            </Form>
+        </div>
+
+        <DeleteUser />
+    </SettingsLayout>
+</template>
